FS-005: reorganize project structure into modules
